Add function to connect to random central server
