--- digs-computational-node.py
@@ -1,13 +1,0 @@
-import argparse
-
-
-def main():
-    parser = argparse.ArgumentParser()
-    # TODO: define command line arguments
-
-    args = parser.parse_args()
-    # TODO: write computational node code
-
-if __name__ == '__main__':
-    main()
-
