+++ digs-worker-node.py
@@ -1,0 +1,22 @@
+import argparse
+
+
+def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        '-c', '--conf', required=False, default="",
+        help="Provide additional configuration file, on top of the default "
+             "ones."
+    )
+    parser.add_argument(
+        '-d', '--debug', action="store_true", default=False,
+        help="Enable debug mode."
+    )
+
+    args = parser.parse_args()
+
+
+
+if __name__ == '__main__':
+    main()
+
